move the test to correct place and format
