+++ hip/test/CMakeLists.txt
@@ -4,7 +4,6 @@ add_subdirectory(base)
 add_subdirectory(components)
 add_subdirectory(factorization)
 add_subdirectory(matrix)
-add_subdirectory(multigrid)
 add_subdirectory(solver)
 add_subdirectory(preconditioner)
 add_subdirectory(stop)
