+++ omp/test/CMakeLists.txt
@@ -3,7 +3,6 @@ include(${PROJECT_SOURCE_DIR}/cmake/create_test.cmake)
 add_subdirectory(components)
 add_subdirectory(factorization)
 add_subdirectory(matrix)
-add_subdirectory(multigrid)
 add_subdirectory(preconditioner)
 add_subdirectory(reorder)
 add_subdirectory(solver)
